s/Focussed/Focused/g. Add lastSelectedNode callbacks and event listener
* examples/recorder.py: s/Focussed/Focused/g. Add lastSelectedNode
callbacks and event listener registration. s/MouseClick/MouseButton/g.  Keep
track of lastPressedNode and lastReleasedNode. Add code to try to detect the
"real" source nodes by checking the X and Y coordinates of lastFocusedNode and
lastSelectedNode. Add a way to exit the recorder - currently, clicking on an
Edit menu will do it. When the OOScriptWriter is initialized, make it spit out
"from dogtail.tree import *".