Add gtk.Expander and gtk.TextView to allow viewing and editing a script.
* dogtail/recorder/recorder.glade: Add gtk.Expander and	gtk.TextView
to allow viewing and editing a script. Note that when the Expander is
collapsed, the main window doesn't shrink back to its original size.  This is a
bug.

* dogtail/recorder/dogtail-recorder: Make ScriptWriters either store
the script in a normal Python string, or a gtk.TextView if it's available.
Continue to print the script to standard out, for debugging. Implement saving
scripts via the GUI, automatically appending a '.py' to the filename if
necessary. Also, add some basic docstrings to some OOScriptWriter methods
(mostly to note that the keyboard stuff doesn't work yet).