Add new Node attributes: 'focusable' and 'focused'. Both are read-only
* dogtail/tree.py: Add new Node attributes: 'focusable' and 'focused'.
Both are read-only Boolean values. 'focusable' indicates whether the Node is
able of taking keyboard focus. 'focused' indicates whether the Node currently
has keyboard focus. Remove Node.rawType() and move its functionality to
Node.typeText(). In Node.typeText(), if the node is focusable and not focused,
try to grab its focus with Node.grabFocus before using rawtype.typeText() to do
the typing. If the node is not focusable, just append the string to the node's
'text' property. Add Node.keyCombo(), which does a similar thing but cannot
have a fallback mechanism.