Make distribution detection code a little less ugly. Remove docstrings
* dogtail/distro.py: Make distribution detection code a little less
ugly. Remove docstrings from the Distro subclasses. Clean up some other
documentation.

* dogtail/i18n.py: Don't import the distro module until we actually
need it.

* dogtail/logging.py: Add an optional 'newline' arg to Logger.log(),
which defaults to True. Passing False will cause the logger to not print or
write a newline at the end of the message. Also stop catching some IOErrors, as
that made the code awkward for no good reason.


svn path=/trunk/; revision=385