Add quiet mode
Normally, the X server logs and other debugging stuff is printed out
when running dogtail-run-headless.  This commit adds a CLI option to
disable all such messages.  This is useful when we want to watch the
dogtail script output in a terminal rather than logging it to a file
with the -o or -e options.