Added two new functions to do a motion with trajectories
A new functions to do drag and absolute motiuon with trajectories were added. The 'trajectory' means that the absolute motion is developed from several partial movements to simulate the motion properly. This also enables reation of some widgets which are less sensitive to motion and the motion needs cast longer.