Add support for only searching among 'showing' nodes.
Here 'showing' refers to pyatspi.STATE_SHOWING, i.e. whether a node is
shown to the end-user or not. Quite often we are only interested in
such nodes, at least when dogtail is used to interact with an
application (e.g. clicking something that isn't there won't
work). Most importantly, this greatly simplifies situations where the
'shown' element we are looking for is hard to exactly pinpoint since
it lacks properties to distinguish it from some not 'shown' elements,
which is quite common when nodes lack names.

Therefore we add a `showingOnly` boolean flag to all Node search
methods. The default will be to not do this, for backwards
compatibility, but the default is configurable via a new
`searchShowingOnly` config option.