Discovered py2 unicode pain within Tree and elsewhere, if we get unicode characters from at-spi and are (now) using the __future unicode_character import.
We're using __future, but that only makes directly declared strings unicode.
Then we get say Node.name as plain str in py2. Should it contain unicode char,
for example '…' an attempt to cat with declared strings that are unicode
(and cannot be otherwise thanks to __future), we're doomed.

Attempted solution: doing plain str() while we use __future seems to make things
click as long as all parts of the concat are str(). Not absolutely sure why - but
on the good side, works the same in both python versions without having to do
.decode() on str strings and having to catch exception for python3.

Sideline: unicode and py2 with py3 - lot's of pain.